requirements.txt
# MISC
PyYAML==6.0

# ASYNC
uvloop==0.17.0

# HTTP
requests==2.31.0
aiohttp>=3.7.4

# SEARCH
fuzzywuzzy[speedup]==0.18.0

# API's
discord.py[voice]==2.2.3
PyGithub==1.58.1

# CHARTS
matplotlib>=3.1.1
pandas>=1.0.0
emoji==2.2.0

# TEXT FORMATTING
titlecase==2.4.0

# Discord.py DEBUG
# jishaku==2.5.1  <-- This relies on line-profiler which doesn't build ATM on Apple Silicon :(
jishaku @ git+https://github.com/Gorialis/jishaku@bf26bea

# TESTING
# tox==3.13.2

# DATABASE
asyncpg==0.27.0
# aiopg==1.0.0

# CACHE
# https://github.com/argaen/aiocache
ujson==5.7.0
msgpack==1.0.5
aiocache[redis,memcached]==0.12.1

# https://github.com/aio-libs
# aioredis==1.2.0
# aiomcache==0.6.0
